Defer showing of dialogs until they're loaded
Currently, dialogs are shown immediately, which incurs a delay with a blank screen while their resources load.

This defers the showing of the About and Preferences dialogs until they're loaded.

In the case of the About dialog, we wait for the built-in `ready-to-show` event, as [recommended by the Electron documentation](https://electronjs.org/docs/api/browser-window#showing-window-gracefully).

The Preferences window, on the other hand, has some extra resources to gather, in this case the icon web font, so instead I've made it wait for the fonts to finish loading, by making use of the `document.fonts.ready` Promise!

Just to be thorough, I've made these two dialogs use the background colour they'll eventually have while loading, just in case they experience any other slow loads. Makes it a bit more seamless 😄