Avoid to access null when the issue author is a ghost
When Jasper fetches an issue which the author is a ghost (deleted user),
Jasper prints a TypeError on GitHubV4IssueClient.ts:51.

```
TypeError: Cannot read property 'login' of null
```

Because GitHub API returns `null` as the author when the user is
deleted.

This patch fixes this problem with adding a null-guard.